+++ lib/commands/update/update_command.dart
@@ -1,20 +1,40 @@
 import 'dart:io';
 import 'package:alex/runner/alex_command.dart';
+import 'package:alex/src/exception/run_exception.dart';
+import 'package:alex/src/local_data.dart';
+import 'package:alex/src/system/update_checker.dart';
 import 'package:alex/src/version.dart';
 
-import 'check_command.dart';
-
 class UpdateCommand extends AlexCommand {
-  UpdateCommand()
-      : super(
+  static const _argChange = CmdArg('check');
+
+  final AlexLocalData _localData;
+  UpdateCommand({AlexLocalData? localData})
+      : _localData = localData ?? AlexLocalData(),
+        super(
           'update',
           'Update alex to the latest version from pub.dev.',
         ) {
-    addSubcommand(UpdateCheckCommand());
+    argParser
+      ..addFlagArg(
+        _argChange,
+        help: 'Run check if there are updates available for alex.',
+      );
   }
 
   @override
   Future<int> doRun() async {
+    final ar = argResults!;
+    final isCheck = ar.getBool(_argChange) || ar.rest.contains('check');
+
+    if (isCheck) {
+      return _runCheck();
+    } else {
+      return _runUpdate();
+    }
+  }
+
+  Future<int> _runUpdate() async {
     printInfo('Updating alex...');
     const currentVersion = packageVersion;
 
@@ -43,6 +63,26 @@ class UpdateCommand extends AlexCommand {
       return success(message: message);
     } else {
       return error(result.exitCode, message: 'Failed to update alex.');
+    }
+  }
+
+  Future<int> _runCheck() async {
+    final checker = UpdateChecker(_localData, out);
+    final result = await checker.run();
+
+    switch (result) {
+      case UpdateCheckResultUpToDate():
+        return success(message: '✅ alex is up to date.');
+      case UpdateCheckResultUpdateAvailable():
+        return success();
+      case UpdateCheckFailure():
+        return error(
+          1,
+          message:
+              "❌ Update check failed. Check your internet connection or try again later.",
+        );
+      case UpdateCheckSkipped():
+        throw const RunException.err('Unexpected result: check was skipped.');
     }
   }